modify : user_index 자료형 수정
user_index 자료형 Long에서 int로

+++ src/main/java/com/bringup/member/membership/domain/entity/UserMembership.java
@@ -12,7 +12,7 @@ public class UserMembership {
 
     @Id
     @Column(name = "user_index", nullable = false)
-    private Long userIndex;
+    private int userIndex;
 
     @Column(name = "start_date", nullable = false, length = 30)
     private String startDate;
